separate tiny_jpeg.h from main.c
